chore(parser): remove legacy sector and BCA parsers
