+++ parser/models/__init__.py
@@ -1,0 +1,11 @@
+from .enums import ClassType, WeekDay, RawWeekDay
+from .subject import Subject
+from .class_info import ClassInfo
+
+__all__ = [
+    "ClassType",
+    "WeekDay",
+    "RawWeekDay",
+    "Subject",
+    "ClassInfo",
+]
