+++ parser/models/class_info.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+from .enums import ClassType
+
+
+class ClassInfo(BaseModel):
+    """Information about a single class."""
+    subject_name: str
+    type: ClassType
+    location: str
