+++ parser/modules/compare_tt/__init__.py
@@ -1,0 +1,3 @@
+from .compare import compare_timetables, _expand_timetable_to_hourly
+
+__all__ = ["compare_timetables", "_expand_timetable_to_hourly"]
